students/k3341/Korchagin_Vadim/Lr1/lab1/app/db/connection.py
import os
from sqlmodel import SQLModel, create_engine, Session
from dotenv import load_dotenv

load_dotenv()

db_url = os.getenv('PR1_DB_URL')
engine = create_engine(db_url, echo=True)

def init_db() -> None:
    SQLModel.metadata.create_all(engine)

def get_session():
    with Session(engine) as session:
        yield session