students/k3341/Korchagin_Vadim/Lr1/requirements.txt
alembic==1.10.2
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.95.0
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.15
passlib==1.7.4
psycopg==3.1.18
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.21
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.8
starlette==0.26.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uvicorn==0.23.2
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
